requirements.txt
certifi==2022.6.15
digit-interface==0.2.1
numpy==1.23.1
opencv-python~=4.5.3.56
pip==22.1.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyudev==0.23.2
setuptools==61.2.0
six==1.16.0
wheel==0.37.1

torch~=1.12.0
pandas~=1.4.3
pillow~=9.2.0
torchvision~=0.13.0
scipy~=1.9.0
open3d~=0.15.2
matplotlib~=3.5.2
attrdict~=2.0.1
hydra~=2.5
imageio~=2.21.0